Move lite credit card input to separate folder + bug fix card image
