+++ src/index.js
@@ -1,6 +1,6 @@
 import connectToState from "./connectToState";
 import CCF from "./CreditCardInput";
-import LiteCCF from "./LiteCreditCardInput";
+import LiteCCF from "./LiteCreditCardInput/LiteCreditCardInput";
 import CV from "./CardView";
 
 export const CreditCardInput = connectToState(CCF);
